--- packages/blocks-editor/README.md
@@ -1,1 +1,0 @@
-# blocks-editor
