--- packages/blocks-editor/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "blocks-editor",
-  "main": "dist/blocks-editor.cjs.js",
-  "module": "dist/blocks-editor.esm.js",
-  "version": "0.0.1",
-  "license": "MIT",
-  "repository": "blocks/blocks-editor",
-  "dependencies": {
-    "@theme-ui/components": "^0.3.1",
-    "@theme-ui/style-guide": "^0.3.0",
-    "color": "^3.1.2",
-    "react-color": "^2.18.0",
-    "reakit": "^1.0.0-beta.14",
-    "tinycolor2": "^1.4.1"
-  },
-  "devDependencies": {
-    "@emotion/core": "^10.0.22",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
-    "theme-ui": "^0.3.1"
-  },
-  "peerDependencies": {
-    "@emotion/core": "^10.0.22",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
-    "theme-ui": "^0.3.1"
-  }
-}
