+++ packages/blocks-ui/src/index.js
@@ -1,9 +1,5 @@
-import * as edit from 'blocks-editor'
-
 export { default } from './editor'
 export * as Controls from './controls'
 export { default as InlineRender } from './inline-render'
 export * as queries from './queries'
 export * as transforms from './transforms'
-
-console.log(edit)
