+++ packages/blocks-ui/src/theme-editor/EditorProvider.js
